Разделяю программу на слои
Теперь программа расслоена на 3 слоя: жанры, основной функционал, api для работы с TMDb, интерфейс.
