+++ movieRoller/movieRoller.csproj
@@ -106,6 +106,9 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
+    <Compile Include="layers\Genres.cs" />
+    <Compile Include="layers\TMDbApi.cs" />
+    <Compile Include="layers\Roller.cs" />
     <Compile Include="MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
